+++ src/pages/Login/LoginPage.sagas.ts
@@ -4,33 +4,33 @@ import { call, put, takeEvery } from "redux-saga/effects";
 import { baseAPIURL } from "../../shared/constants";
 import { LocalStorage } from "../../app/common/LocalStorage";
 import {
-  requestLogin,
-  receiveLoginFailure,
-  receiveLoginSuccess,
-  IRequestLoginAction,
+	requestLogin,
+	receiveLoginFailure,
+	receiveLoginSuccess,
+	IRequestLoginAction,
 } from "./LoginPage.slice";
 
 function* handleRequestLogin(action: PayloadAction<IRequestLoginAction>): any {
-  try {
-    const url = `${baseAPIURL}v1/user/login`;
-    const { data } = yield call(async () => {
-      return axios.post(url, { data: action.payload });
-    });
-    const { token } = data;
-    LocalStorage.setToken(token);
-    yield put(receiveLoginSuccess(token));
-  } catch (e) {
-    LocalStorage.checkUnauthorized(e);
-    yield put(
-      receiveLoginFailure(
-        "The username or password is incorrect. Please try again."
-      )
-    );
-  }
+	try {
+		const url = `${baseAPIURL}v1/user/login`;
+		const { data } = yield call(async () => {
+			return axios.post(url, action.payload);
+		});
+		const { token } = data;
+		LocalStorage.setToken(token);
+		yield put(receiveLoginSuccess(token));
+	} catch (e) {
+		LocalStorage.checkUnauthorized(e);
+		yield put(
+			receiveLoginFailure(
+				"The username or password is incorrect. Please try again."
+			)
+		);
+	}
 }
 
 function* loginSaga() {
-  yield takeEvery(requestLogin.type, handleRequestLogin);
+	yield takeEvery(requestLogin.type, handleRequestLogin);
 }
 
 export default loginSaga;
