+++ tsconfig.json
@@ -16,5 +16,5 @@
 		"noEmit": true,
 		"jsx": "react-jsx"
 	},
-	"include": ["src", "src/__tests__", "auth", "src/__tests__"]
+	"include": ["src"]
 }
